refactor: Activity 클래스들을 activity 패키지 밖으로 이동
