+++ app/src/main/java/woowacourse/shopping/ui/shopping/ProductListActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.shopping.activity
+package woowacourse.shopping.ui.shopping
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,8 +7,8 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import woowacourse.shopping.ProductFixture
 import woowacourse.shopping.domain.Products
-import woowacourse.shopping.ui.cart.activity.CartActivity
-import woowacourse.shopping.ui.productdetail.activity.ProductDetailActivity
+import woowacourse.shopping.ui.cart.CartActivity
+import woowacourse.shopping.ui.productdetail.ProductDetailActivity
 import woowacourse.shopping.ui.shopping.screen.ProductListScreen
 import woowacourse.shopping.ui.theme.androidshoppingTheme
